requirements.txt
SQLAlchemy==1.4.37
requests==2.27.1
pdfplumber==0.7.6
python-dotenv==0.19.2
asyncio==3.4.3
python-dateutil==2.8.2
black~=23.3.0
xlrd >= 2.0.1
pandas==2.3.2
